+++ compiler-core/src/erlang.rs
@@ -129,7 +129,7 @@ pub fn record_definition(name: &str, fields: &[(&str, Arc<Type>)]) -> String {
         docvec!(atom_string((*name).to_string()), " :: ", type_.group())
     });
     let fields = break_("", "")
-        .append(concat(Itertools::intersperse(fields, break_(",", ", "))))
+        .append(join(fields, break_(",", ", ")))
         .nest(INDENT)
         .append(break_("", ""))
         .group();
@@ -176,36 +176,24 @@ fn module_document<'a>(
         (false, false) => return Ok(header),
         (true, false) => "-export(["
             .to_doc()
-            .append(concat(Itertools::intersperse(
-                exports.into_iter(),
-                ", ".to_doc(),
-            )))
+            .append(join(exports, ", ".to_doc()))
             .append("]).")
             .append(lines(2)),
 
         (true, true) => "-export(["
             .to_doc()
-            .append(concat(Itertools::intersperse(
-                exports.into_iter(),
-                ", ".to_doc(),
-            )))
+            .append(join(exports, ", ".to_doc()))
             .append("]).")
             .append(line())
             .append("-export_type([")
             .to_doc()
-            .append(concat(Itertools::intersperse(
-                type_exports.into_iter(),
-                ", ".to_doc(),
-            )))
+            .append(join(type_exports, ", ".to_doc()))
             .append("]).")
             .append(lines(2)),
 
         (false, true) => "-export_type(["
             .to_doc()
-            .append(concat(Itertools::intersperse(
-                type_exports.into_iter(),
-                ", ".to_doc(),
-            )))
+            .append(join(type_exports, ", ".to_doc()))
             .append("]).")
             .append(lines(2)),
     };
@@ -213,16 +201,16 @@ fn module_document<'a>(
     let type_defs = if type_defs.is_empty() {
         nil()
     } else {
-        concat(Itertools::intersperse(type_defs.into_iter(), lines(2))).append(lines(2))
+        join(type_defs, lines(2)).append(lines(2))
     };
 
-    let statements = concat(Itertools::intersperse(
+    let statements = join(
         module
             .definitions
             .iter()
             .flat_map(|s| module_statement(s, &module.name, line_numbers)),
         lines(2),
-    ));
+    );
 
     Ok(header
         .append("-compile([no_auto_import, nowarn_unused_vars, nowarn_unused_function, nowarn_nomatch]).")
@@ -299,7 +287,7 @@ fn register_imports(
             let definition = if constructors.is_empty() {
                 let constructors =
                     std::iter::once("any()".to_doc()).chain(phantom_vars_constructor);
-                concat(Itertools::intersperse(constructors, break_(" |", " | ")))
+                join(constructors, break_(" |", " | "))
             } else {
                 let constructors = constructors
                     .iter()
@@ -314,14 +302,14 @@ fn register_imports(
                         }
                     })
                     .chain(phantom_vars_constructor);
-                concat(Itertools::intersperse(constructors, break_(" |", " | ")))
+                join(constructors, break_(" |", " | "))
             }
             .nest(INDENT);
             let type_printer = TypePrinter::new(module_name);
-            let params = concat(Itertools::intersperse(
+            let params = join(
                 typed_parameters.iter().map(|a| type_printer.print(a)),
                 ", ".to_doc(),
-            ));
+            );
             let doc = if *opaque { "-opaque " } else { "-type " }
                 .to_doc()
                 .append(Document::String(erl_safe_type_name(name.to_snake_case())))
@@ -416,10 +404,7 @@ where
     I: IntoIterator<Item = Document<'a>>,
 {
     break_("", "")
-        .append(concat(Itertools::intersperse(
-            args.into_iter(),
-            break_(",", ", "),
-        )))
+        .append(join(args, break_(",", ", ")))
         .nest(INDENT)
         .append(break_("", ""))
         .surround("(", ")")
@@ -507,7 +492,7 @@ fn string(value: &str) -> Document<'_> {
 }
 
 fn tuple<'a>(elems: impl IntoIterator<Item = Document<'a>>) -> Document<'a> {
-    concat(Itertools::intersperse(elems.into_iter(), break_(",", ", ")))
+    join(elems, break_(",", ", "))
         .nest(INDENT)
         .surround("{", "}")
         .group()
@@ -559,7 +544,7 @@ fn string_concatenate_argument<'a>(value: &'a TypedExpr, env: &mut Env<'a>) -> D
 }
 
 fn bit_array<'a>(elems: impl IntoIterator<Item = Document<'a>>) -> Document<'a> {
-    concat(Itertools::intersperse(elems.into_iter(), break_(",", ", ")))
+    join(elems, break_(",", ", "))
         .nest(INDENT)
         .surround("<<", ">>")
         .group()
@@ -928,18 +913,20 @@ fn expr_list<'a>(
     tail: &'a Option<Box<TypedExpr>>,
     env: &mut Env<'a>,
 ) -> Document<'a> {
-    let elements = concat(Itertools::intersperse(
+    let elements = join(
         elements.iter().map(|e| maybe_block_expr(e, env)),
         break_(",", ", "),
-    ));
+    );
     list(elements, tail.as_ref().map(|e| maybe_block_expr(e, env)))
 }
 
 fn list<'a>(elems: Document<'a>, tail: Option<Document<'a>>) -> Document<'a> {
-    let elems = if let Some(final_tail) = tail {
-        elems.append(break_(" |", " | ")).append(final_tail)
-    } else {
-        elems
+    let elems = match tail {
+        Some(tail) if elems.is_empty() => return tail.to_doc(),
+
+        Some(tail) => elems.append(break_(" |", " | ")).append(tail),
+
+        None => elems,
     };
 
     elems.to_doc().nest(INDENT).surround("[", "]").group()
@@ -1012,13 +999,13 @@ fn const_inline<'a>(literal: &'a TypedConstant, env: &mut Env<'a>) -> Document<'
         Constant::String { value, .. } => string(value),
         Constant::Tuple { elements, .. } => tuple(elements.iter().map(|e| const_inline(e, env))),
 
-        Constant::List { elements, .. } => {
-            let elements = Itertools::intersperse(
-                elements.iter().map(|e| const_inline(e, env)),
-                break_(",", ", "),
-            );
-            concat(elements).nest(INDENT).surround("[", "]").group()
-        }
+        Constant::List { elements, .. } => join(
+            elements.iter().map(|e| const_inline(e, env)),
+            break_(",", ", "),
+        )
+        .nest(INDENT)
+        .surround("[", "]")
+        .group(),
 
         Constant::BitArray { segments, .. } => bit_array(
             segments
@@ -1078,7 +1065,7 @@ fn clause<'a>(clause: &'a TypedClause, env: &mut Env<'a>) -> Document<'a> {
     let initial_erlang_vars = env.erl_function_scope_vars.clone();
     let mut end_erlang_vars = im::HashMap::new();
 
-    let docs = Itertools::intersperse(
+    let doc = join(
         std::iter::once(pat)
             .chain(alternative_patterns)
             .map(|patterns| {
@@ -1106,7 +1093,6 @@ fn clause<'a>(clause: &'a TypedClause, env: &mut Env<'a>) -> Document<'a> {
         ";".to_doc().append(lines(2)),
     );
 
-    let doc = concat(docs);
     env.erl_function_scope_vars = end_erlang_vars;
     doc
 }
@@ -1237,7 +1223,7 @@ fn clause_guard<'a>(guard: &'a TypedClauseGuard, env: &mut Env<'a>) -> Document<
 }
 
 fn clauses<'a>(cs: &'a [TypedClause], env: &mut Env<'a>) -> Document<'a> {
-    concat(Itertools::intersperse(
+    join(
         cs.iter().map(|c| {
             let vars = env.current_scope_vars.clone();
             let erl = clause(c, env);
@@ -1245,7 +1231,7 @@ fn clauses<'a>(cs: &'a [TypedClause], env: &mut Env<'a>) -> Document<'a> {
             erl
         }),
         ";".to_doc().append(lines(2)),
-    ))
+    )
 }
 
 fn case<'a>(subjects: &'a [TypedExpr], cs: &'a [TypedClause], env: &mut Env<'a>) -> Document<'a> {
@@ -1999,10 +1985,7 @@ impl<'a> TypePrinter<'a> {
     }
 
     fn print_type_app(&self, module: &str, name: &str, args: &[Arc<Type>]) -> Document<'static> {
-        let args = concat(Itertools::intersperse(
-            args.iter().map(|a| self.print(a)),
-            ", ".to_doc(),
-        ));
+        let args = join(args.iter().map(|a| self.print(a)), ", ".to_doc());
         let name = Document::String(erl_safe_type_name(name.to_snake_case()));
         if self.current_module == module {
             docvec![name, "(", args, ")"]
@@ -2012,10 +1995,7 @@ impl<'a> TypePrinter<'a> {
     }
 
     fn print_fn(&self, args: &[Arc<Type>], retrn: &Type) -> Document<'static> {
-        let args = concat(Itertools::intersperse(
-            args.iter().map(|a| self.print(a)),
-            ", ".to_doc(),
-        ));
+        let args = join(args.iter().map(|a| self.print(a)), ", ".to_doc());
         let retrn = self.print(retrn);
         "fun(("
             .to_doc()